fix(preprocessor): pipe stdout to a user's running interface to show warning messages from a subprocess
This is a dual-purpose fix. (1) The Preprocessor used to print out stderr, instead of raising an exception. The change makes python aware of an incomplete process and properly raise a CalledProcessError, which will be handled by the Exceptions; (2) 'stdout=subprocess.PIPE' prints out warning messages from bcftools and helps diagnose subtle VCF issues